feat: separate pazuz and lamassu routing tree fix: routes.js routing tree getter
